# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Management API"
  description: "Used to manage the everyday functionalities of a mechanic shop"
  version: "0.0.1"
host: "my-cool-mechanic.onrender.com" 
schemes:
  - "https"
consumes:
  - "application/json" 
produces:
  - "application/json" 
securityDefinitions: 
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /customers:
    post:
      tags:
        - Customer
      summary: "Register Customer"
      description: "Send in customer info, to create a customer profile that will be stored in the database."
      parameters:
        - in: body
          name: body
          description: "New Customer Information"
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"
          
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              phone: "813-777-3993"
              address: "2752 Conch st."
              email: "laurne@gmail.com"
              id: 0
              first_name: "Thomas"
              last_name: "Smith"

    get: 
      tags:
        - Customer
      summary: "Returns all customers"
      description: "Endpoint to recieve an array of customer objects."
      responses:
        200:
          description: "Successfully Retrieved Customers"
          schema:
            $ref: "#/definitions/CustomersResponse"   
          examples:
            application/json:
              phone: "813-777-3993"
              address: "2752 Conch st."
              email: "laurne@gmail.com"
              id: 0
              first_name: "Thomas"
              last_name: "Smith"

    put: 
      tags:
        - Customer
      summary: "Update Customer Endpoint"
      description: "Endpoint to update a customer, requires a token, customer will be identified by the token"
      security:
        - bearerAuth: [] 
      parameters:
        - in: "body"
          name: "body"
          description: "Customer Updates"
          required: true
          schema:
            $ref: "#/definitions/CustomerInput"

      responses:
        200:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              phone: "813-777-3993"
              address: "2752 Conch st."
              email: "laurne@gmail.com"
              id: 0
              first_name: "Thomas"
              last_name: "Smith"

    delete: 
      tags:
        - Customer
      summary: "Delete a Customer"
      description: "Delete a Customer, requires a token, token will be used to identify customer to delete."
      security:
        - bearerAuth: []

      responses:
        200:
          schema:
            type: object
            properties:
              message:
                type: string
  /customers/profile:
    get:
      tags:
        - Customer
      summary: "Returns the profile of the authorized customer"
      description: "Endpoint retrieves the profile of the customer identified via the token."
      security:
        - bearerAuth: []   # If you use token auth
      responses:
        200:
          description: "Successfully retrieved customer profile"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              phone: "813-777-3993"
              address: "2752 Conch st."
              email: "lauren@gmail.com"
              id: 0
              first_name: "Thomas"    
              last_name: "Smith"
  /customers/paginated:
    get:
      tags:
        - Customer
      summary: "Get a paginated list of customers"
      description: "Endpoint to recieve an array of customer objects in a paginated format."
      parameters:
        - in: query
          name: page
          type: integer
          required: false
          description: "Page number (default is 1)"
        - in: query
          name: per_page
          type: integer
          required: false
          description: "Number of items per page (default is 10)"
      responses:
        200:
          description: "Successfully Retrieved Paginated Customers"
          schema:
            $ref: "#/definitions/PaginatedCustomersResponse"
          examples:
            application/json:
              total: 50
              pages: 5
              current_page: 1
              per_page: 10
              customers:
                - phone: "813-777-3993"
                  address: "2752 Conch st."
                  email: "lauren@gmail.com"
                  id: 0
                  first_name: "Thomas"
                  last_name: "Smith"  
...
definitions: 
...
  CustomerInput: 
    type: object
    properties:
      phone:
        type: string
        example: "813-777-3993"
      address:
        type: string
        example: "2752 Conch St."
      email:
        type: string
        example: "lauren@gmail.com"
      first_name:
        type: string
        example: "Thomas"
      last_name:
        type: string
        example: "Smith" 

    required:
      - first_name
      - last_name
      - email
      - phone
      - address
  CustomerResponse: 
    type: object
    properties:
      phone:
        type: string
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
  CustomersResponse: 
    type: array
    items:
      type: object
      properties:
        phone:
          type: string
        address:
          type: string
        email:
          type: string
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
  PaginatedCustomersResponse:
    type: object      
    properties:
      total:
        type: integer
      pages:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      customers:
        type: array
        items:
          $ref: "#/definitions/CustomerResponse"
...